# [DC-420] Multithread storage system dataset fetching