Before:
curl -X 'GET' 'localhost:8080/api/v1/datasets' -H 'accept: application/json'   0.01s user 0.01s system 1% cpu 1.451 total

After:
curl -X 'GET' 'localhost:8080/api/v1/datasets' -H 'accept: application/json'   0.01s user 0.02s system 7% cpu 0.292 total

This is round one of performance improvements. The next planned change is to do all the database fetching at once, rather than do a separate dao call per storage system.